- refactor react-force-graph-2d module augmentation to react-force-graph-2d.d.ts file
- refactor types/utilities to supporting files per above change
