+++ website/client/components/GraphView/EntryInfoItem.tsx
@@ -12,7 +12,8 @@ import TableRow from '@mui/material/TableRow';
 import Collapse from '@mui/material/Collapse';
 import ExpandLess from '@mui/icons-material/ExpandLess';
 import ExpandMore from '@mui/icons-material/ExpandMore';
-import {EntryInfo, displayNames } from "@components/GraphView/util";
+import { displayNames } from "@components/GraphView/util";
+import {EntryInfo } from "@components/GraphView/GraphTypes";
 import SearchIcon from '@mui/icons-material/Search';
 import ListItemIcon from "@mui/material/ListItemIcon";
 import Tooltip from "@mui/material/Tooltip";
